Load 3 more fonts
1) Roboto
2) Font Awesome 5 Regular
3) Font Awesome 5 Solid

They are all merged so they can be used all in the same string

See instructions from DearImGui to generate the header file containing the font. You need to use binary_to_compressed_c.cpp.
Once the header file is generated, it can be included and used as a font.
The icons are in the font from Font Awesome. The helper from IconFontCppHeaders provides a map of icon name to its corresponding entry in the font.